prompt/ai-audit-comments/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-APScheduler==1.13.1
python-dotenv==1.1.0
openai==1.73.0
